pyproject.toml
[tool.poetry]
name = "repl_python3_Flask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Pillow = "^9.1.0"
Pandas3 = "^0.0.1"
waitress = "^2.1.1"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
